Moved base58check to its own package to avoid adding base58check.go to go run commands and separate out helper package base58check
